% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHelpFiles.R
\name{getAllHelpControlListParams}
\alias{getAllHelpControlListParams}
\title{getAllHelpControlListParams}
\usage{
getAllHelpControlListParams(methodName, directSpotControl = NULL)
}
\arguments{
\item{methodName}{Name of the method of which the help file should be read}

\item{directSpotControl}{Default NULL to read any other control list. If this isn't NULL, then methodName
is ignored and this parameter is handled as an rd list instead}
}
\value{
A list of all parameters in the control list
}
\description{
Extracts all control list parameters of a given method in the SPOT-package by parsing it's help file.
}
\examples{
spotGUI:::getAllHelpControlListParams("buildKriging")

}
\keyword{internal}
