% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHelpFiles.R
\name{getHelpSpotControlParameters}
\alias{getHelpSpotControlParameters}
\title{getHelpSpotControlParameters}
\usage{
getHelpSpotControlParameters()
}
\value{
A list of all parameters in the control list
}
\description{
Extracts all Parameters specified in spotControl
}
\examples{
spotGUI:::getMethodHelpArgument("SPOT","buildKriging")

}
\keyword{internal}
