% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHelpFiles.R
\name{getMethodHelp}
\alias{getMethodHelp}
\title{getMethodHelp}
\usage{
getMethodHelp(packageName, methodName, getAll = FALSE)
}
\arguments{
\item{packageName}{Name of the package to extract the help files from}

\item{methodName}{Name of the method of which the help file should be read}

\item{getAll}{True if instead of a single one, all help-files of a given package should be read. Default = False}
}
\value{
An Rd Object with the package help file or a list of all Rd-Objects for a given package
}
\description{
Extracts help-files from a given package
}
\examples{
spotGUI:::getMethodHelp("SPOT","buildKriging")
spotGUI:::getMethodHelp("SPOT",NULL, TRUE)

}
\keyword{internal}
