% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHelpFiles.R
\name{getMethodHelpArgument}
\alias{getMethodHelpArgument}
\title{getMethodHelpArgument}
\usage{
getMethodHelpArgument(packageName, methodName)
}
\arguments{
\item{packageName}{Name of the package to extract the help file from}

\item{methodName}{Name of the method of which the help file should be read}
}
\value{
An Rd Object with the "Arguments" section of a given help file. If no "Arguments"
Section exists, this returns NULL
}
\description{
Extracts the "Arguments" section from a help file
}
\examples{
spotGUI:::getMethodHelpArgument("SPOT","buildKriging")

}
\keyword{internal}
