% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHelpFiles.R
\name{getMethodHelpControlList}
\alias{getMethodHelpControlList}
\title{getMethodHelpControlList}
\usage{
getMethodHelpControlList(methodName)
}
\arguments{
\item{methodName}{Name of the method of which the help file should be read}
}
\value{
An Rd Object with the "control" section of a given help file. If no control list is found this returns NULL
}
\description{
Extracts the description of a control list of a specified method in the SPOT package.
}
\examples{
spotGUI:::getMethodHelpControlList("buildKriging")

}
\keyword{internal}
