% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogCreator.R
\name{getTextoutputBestSolution}
\alias{getTextoutputBestSolution}
\title{Textoutput Field 'Best Solution'}
\usage{
getTextoutputBestSolution(input, data)
}
\arguments{
\item{input}{shiny input}

\item{data}{data.frame with all candidate solutions}
}
\value{
ouputField
}
\description{
Generates the outputField to show the best till then found candidate solution
}
