% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUi.R
\name{getUIPage}
\alias{getUIPage}
\title{Define UI of SPOT-GUI}
\usage{
getUIPage()
}
\description{
Generates the UI part of the SPOT-GUI.
This method is used internally in the starting process of the GUI.
Manual use of this function is not advised.
}
