% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogCreator.R
\docType{data}
\name{LIST_ITEMS_PER_ROW}
\alias{LIST_ITEMS_PER_ROW}
\title{List Items per row in R-Code Exports}
\format{An object of class \code{numeric} of length 1.}
\usage{
LIST_ITEMS_PER_ROW
}
\description{
Specify how many items of a list a printed per line
}
\keyword{internal}
