% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogCreator.R
\name{rLogResetMessage}
\alias{rLogResetMessage}
\title{Generate the Initial Code Block for the R-Export}
\usage{
rLogResetMessage()
}
\value{
String, rExport initial code
}
\description{
Generate the Initial Code Block for the R-Export
}
\keyword{internal}
