% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSpotGUI.R
\name{runSpotGUI}
\alias{runSpotGUI}
\title{runSpotGUI}
\usage{
runSpotGUI()
}
\description{
Run the starting command of the SPOT-GUI. Opens the graphical shiny application through which the user
can acess the SPO Toolbox.
}
