% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlDynamicUI.R
\name{xmlGetAllConfiguredControlElements}
\alias{xmlGetAllConfiguredControlElements}
\title{xmlGetAllConfiguredControlElements}
\usage{
xmlGetAllConfiguredControlElements()
}
\value{
List with all configured optimizers, designGenerators and modelBuilders,
as well as their respective parameters.
}
\description{
Parses the config.xml file. Reads all entries for optimizers, designGenerators and modelBuilders.
For each configured entry, a list with the entries name and all of its configurable
parameters is returned.
}
\keyword{internal}
