% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlDynamicUI.R
\name{xmlGetRootElement}
\alias{xmlGetRootElement}
\title{Get root element of the config.xml file of this package}
\usage{
xmlGetRootElement()
}
\value{
list of xml elements
}
\description{
Get root element of the config.xml file of this package
}
\keyword{internal}
